+++ src/UniqueGmailAddress.php
@@ -2,10 +2,10 @@
 
 namespace ImLiam\UniqueGmailAddress;
 
-use ImLiam\UniqueGmailAddress\Rules\DotsInUsername;
+use ImLiam\UniqueGmailAddress\Rules\DotsInLocalPart;
 use ImLiam\UniqueGmailAddress\Rules\GmailDomain;
 use ImLiam\UniqueGmailAddress\Rules\MixedCase;
-use ImLiam\UniqueGmailAddress\Rules\TagInUsername;
+use ImLiam\UniqueGmailAddress\Rules\TagInLocalPart;
 
 class UniqueGmailAddress extends UniqueEmailAddress
 {
@@ -16,8 +16,8 @@ class UniqueGmailAddress extends UniqueEmailAddress
         if (Util::isGmailAddress($this->email) || Util::isGsuiteAddress($this->email)) {
             $this->rules[GmailDomain::class] = new GmailDomain();
             // $this->rules[MixedCase::class] = new MixedCase();
-            $this->rules[DotsInUsername::class] = new DotsInUsername();
-            $this->rules[TagInUsername::class] = new TagInUsername('+');
+            $this->rules[DotsInLocalPart::class] = new DotsInLocalPart();
+            $this->rules[TagInLocalPart::class] = new TagInLocalPart('+');
         }
     }
 }
